% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper functions.R
\name{extract_element}
\alias{extract_element}
\title{Dynamically extract the A matrix from the output of deconvolution function based on a user-specified address}
\usage{
extract_element(data, address)
}
\description{
Dynamically extract the A matrix from the output of deconvolution function based on a user-specified address
}
