% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution_function_RefFreeEWAS.R
\name{RefFreeEwasModel}
\alias{RefFreeEwasModel}
\title{RefFreeEwas package is no longer available online. Here we provide the R scripts for users' reference.}
\usage{
RefFreeEwasModel(Y, X, K, smallOutput = FALSE)
}
\description{
RefFreeEwas package is no longer available online. Here we provide the R scripts for users' reference.
}
