% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_expression_profile}
\alias{sample_expression_profile}
\title{Sample Expression Profile}
\format{
\subsection{\code{sample_expression_profile}}{

A data frame with 30 rows and 5 columns:
}
}
\usage{
sample_expression_profile
}
\description{
The sample-wise cell-type proportion matrix is simulated by the Dirichlet distribution function. We set α (the parameter
controlling the shape of the distribution) at 3 for all cell types in our simulation to make sure the proportions among samples
are basically concentrated towards the center and mutually symmetric. Here we provide an example A matrix for users’ reference,
with 30 samples and 5 cell types.
}
\keyword{datasets}
