% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper functions.R
\name{select_most_variant}
\alias{select_most_variant}
\title{Helper function specifically helping deconvolution methods that require feature selection based on variance ranking}
\usage{
select_most_variant(input, k)
}
\description{
Helper function specifically helping deconvolution methods that require feature selection based on variance ranking
}
